Add new Tabbing stragety for root blocks.
By default, MathQuill allows the browser to handle Tab in the standard way
when the cursor is in the root block.  This means that the user will move to
the next focusable element in tab order.  This behavior bypasses any templates
between the current cursor position and the end.

If templates are thought of as editable areas, like mini-textfields, this
approach feels unexpected to keyboard users, who would expect 'focus' to
traverse to the next 'field,' i.e. next template.

This commit adds a new configuration property: tabAction.  By default tabAction
simply uses the current behavior.  Users can specify a value of
'escapeAndTemplate' to allow Tabs from the root block to find the next template
or to move outside of the MathQuill field if no templates exist.