Initial commit of :before box for unbased exponents
This changes add several new CSS properties built on the new class
mq-empty-base.  mq-empty-base is added to the mq-supsub block when the
preceding block does not exist or contains an operator.

The :before block creates an empty template box, using the
emtpy-template LESS mixin.  The empty-template is also hidden when the
cursor is “inside” of it, which is the same behavior that normal empty
templates use.  This relies on the immediate sibling selector to alter
the background of the template to be transparent.

The README has been updated with configuration information for turning
this on.  Options is tested for supSubsShowEmptyBase to determine if
the mq-empty-base class should be added.

Problems with this commit:
  - Options are not working correctly.  `this` does not have an options
object or controller to obtain one.  In my initial workings,
`this.parent` did, but in running the visual tests that fails as not
all parent’s have controllers.  Need to understand how to get the
options into the method.  Note opts is oftentimes undefined as the
function is called without arguments when used as `respace`.
  - I was unable to get unbased boxes to work correctly for solo
subscripts or for super/sub combinations.
    * I made some efforts using a mq-sub-only class and applying that.
It was mostly correct for the solo case, but not quite right.
    * I tried a myriad of approaches for the combination case, but
nothing worked.

Checking this in to get feedback on the current approach and
suggestions for how to tackle the subscript problems.