Add bower.json file for supporting bower.
The bower version that this project uses is the base version of the release,
the MathQuill release number in package.json, plus a revision number.  Our
revision number is continually increasing, such that 0.9.4-5 -> 0.10.1-6 in
the next release.  This allows us to track the main revision of MathQuill
while also keeping track of our additional features that have not been
accepted via pull request.

We should never submit a pull request to MathQuill for bower.json.