Moved setSelection and clearSelection to AbstractMath.
When updating to the 0.10 release, I misresolved a conflict that led me to put
the setSelection method next to the clearSelection method.  While this seemed
correct at the time, it was not.  The problem with keeping the methods in
EditableMath is that you cannot use them to create mouse selections in
StaticMath.  The select method, which is cursor-related cannot be used in
the StaticMath fields and therefore was not moved.

Updated the documentation to reflect the change in the movement of the API
methods.